// mediumclone/src/app/shared/types/articles.interface.ts
import { ProfileInterface } from "./profile.interface";

export interface ArticlesInterface {
  slug: string;
  title: string;
  description: string;
  body: string;
  tagList: Array<string>;
  createdAt: string;
  updatedAt: string;
  favorited: boolean;
  favoritesCount: number;
  author: ProfileInterface
}